[FIX] Fix NMT request frame size
The nmtcnu_sendNmtRequestEx function is used to send NMT requests to the
MN to invoke an NMT command.
For instance, an RMN sends an NMTGoToStandby request to manually
overtake the AMN status. This commit fixes that the frame size gets
truncated to the minimal NMT request frame size, thus not being able to
transmit any NMT command data.

Change-Id: If3b7312252a26ebea086a214060f912600f8c08c