[FIX] Fix to support removal of multiple VEth adapters
With the current implementation the notify object does not handle
removal of multiple VEth adapters created during the installation
on a system with multiple NICs.

This creates residual VEth adapters and leads to system crashes
during a new installation.

This commit enables the removal of all the VEth adapters created
during the installation using the registry entries created for the
adapters.

Change-Id: Ife8f47c6e27d362b608199ab059e12186154a45c
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>