[FIX] Fix NMT state update in PRes Tx buffer for PreOp2
A CN is only allowed to transmit PRes frames in PreOp2, ReadyToOp and
Operational state. Therefore, it makes no sense to initialize the PRes
Tx buffers to PreOp1 NMT state.
This change makes sure that the PRes Tx buffers are initialized to
PreOp2 NMT state for sudden changes from PreOp1 to PreOp2.

Change-Id: I753b827ec39823ba89231ea9f07614b94162a086