[FIX] Fix Edrv maximum buffer size
The edrv_allocTxBuffer function misses to set the maxBufferSize
parameter of the EdrvTxBuffer structure. Without the assignment in
edrv_allocTxBuffer maxBufferSize is set to the size of the frame. With
this fix commit it is reverted to the value of EDRV_MAX_FRAME_SIZE as it
used to be before implementation of the bufalloc library.

Change-Id: Iab7d80f458cb052526fad935ad7d04da1ce18435