[FIX] Prevent multiple call to string object callback
Due to the change of the SDO to OD connection, the OD gets now
calls for all SDO segments in case of a segmented transfers
to acquire the object pointer. The original CFMU O-string object
reallocation callback function cfmu_cbObdAccess() was designed
to get a call only at the beginning of the object write access.
Therefore, subsequent calls to object callbacks which use the
event kObdEvWrStringDomain (e.g. cfmu_cbObdAccess()) are now
prevented for segmented SDO transfers.
For this purpose, the code part enclosed in the macro
CONFIG_OBD_USE_STRING_DOMAIN_IN_RAM was moved to a new function
reallocStringDomainObj().

Change-Id: Icebf57ca7adc033c60691c55661efc6140113f98