MN Objdict: add new dynamic channel for REAL PDO objects
From [1]:
The OC 2.0.0 is a bit stricter than the former versions in terms of how the MN
PI is auto generated. Therefore it checks for a dynamic channel supporting the
Real32 datatype which is not existing in the MN XDD shipped with the stack.

So add it with a MaxNumber set to 2032 in order to use env 400 REAL objects.

[1] https://sourceforge.net/p/openconf/discussion/help/thread/29fffde5/#54ce

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
---
v2:
Following the review on Github [1], the new PDO channel must
be added to the processImageLink_l[] tab.

[1] https://github.com/OpenAutomationTechnologies/openPOWERLINK_V2/pull/111