[FIX] Fix unconstrained path/port warnings on DE2-115 MN and CN designs
Due to missing constraints for the LED GPIOs, NODE_SWITCH and HEX in
Terasic DE2-115 mn-single-hostif-drv, cn-single-hostif-drv and
cn-dual-hostif-gpio designs, unconstrained paths/ports warnings are
reported in the TimeQuest Timing Analyzer.

This commit add the constraints for the LED GPIOs, NODE_SWITCH and HEX
for Terasic DE2-115 designs which resolves the unconstrained path/ports
warning during placement and routing.

Change-Id: I28f408811e06e53fe22f7f1c1db76d3661bf218c
Signed-off-by: Sudharsan Sivakumar <sudharsan.s@kalycito.com>