[TASK] Replace clock crossing by pipeline bridge
The clock crossing bridge is optimized for issuing multiple access
requests to the Avalon MM slaves at once without halting the Avalon
MM master. This introduces latency and increases resource utilization
unnecessarily for Avalon MM slaves which are not accessed burst-wise.

Therefore the clock crossing bridge is replaced with a pipeline bridge
which is configured to require no additional resources and introduces
no extra delays. Qsys adds a clock crossing adapter to accommodate
with the different clock domains of the MM master and slaves.

Change-Id: I6635b8181c1c8c2752b60fa8a075cb883d7d8b23