[FIX] Correct SDO error code for non-existing objects
Instead of a general error, the following correct abort
code is returned:
* Object does not exist in the object dictionary

The bug was caused by a blocked access to the user specific
object dictionary. Per default, there is no handling implemented
for the user specific OD, and therefore the finish function is never
called. That's why the second access to a non-existing object was
blocked, and returned a wrong SDO abort code.

Change-Id: Ifa77985f4874876b0a5e0061f5303a148149bff7