[FIX] veth: avoid kernel header issue with musl
The Virtual Ethernet driver doesn't build when the musl libc is used on the
system. As stated in the musl wiki [1], the userspace and kernel headers are
mixed leading to a "clash" with the definitions provided by musl.

Remove netinet/if_ether.h userspace header and replace ETHER_ADDR_LEN by
ETH_ALEN [2] and ETHERMTU by ETH_DATA_LEN [3] in veth-linuxuser.c.

Fixes:
http://autobuild.buildroot.org/results/2ca/2ca04bb046263e479e7597867b56469893d3c11d/build-end.log

[1] http://wiki.musl-libc.org/wiki/FAQ#Q:_why_am_i_getting_.22error:_redefinition_of_struct_ethhdr.2Ftcphdr.2Fetc.22_.3F
[2] https://git.musl-libc.org/cgit/musl/tree/include/net/ethernet.h?h=v1.1.14#n35
[3] https://git.musl-libc.org/cgit/musl/tree/include/net/ethernet.h?h=v1.1.14#n48

Change-Id: I4d8008b1f12e8238c06c92a43490d26a314a65b5
Signed-off-by: Romain Naour <romain.naour@gmail.com>