[FIX] Enable storage attribute for PDO comm and mapp param objects
CN console demo using the configuration store/restore feature
on OS based platforms fails to restore PDO mapping configuration.
This occurs as the PDO configurations are not saved to the
non-volatile memory i.e. the disk, due to absence of the storage
attribute for required objects.

This commit enables the storage attribute for PDO communication
and mapping objects allowing them to be stored on disk.

Based on fix in v1.8.x by
        Daniel Krueger <daniel.krueger@systec-electronic.com>

Change-Id: I3d6aa1146ad3d0abde29150e9e16c8f38add565d
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>