[FIX] Force 0x1F8D/F0 object in Demo_3CN openCONFIGURATOR project
If the CNs in the Demo_3CN openCONFIGURATOR project are configured as
chained stations, the CNs report a PDO mapping error. This is due to
the wrong configuration of NMT_PresPayloadLimitList_AU16 (0x1F8D/F0)
object at the PRC CNs.

This fix ensures the correct configuration of the CNs by forcing the
actual value of NMT_PresPayloadLimitList_AU16 (0x1F8D/F0) written to
the CDC file.

Change-Id: I090145fa13a737f948f8f411eaff276588d9e474
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>