[FIX] Improve robustness of SDO sequence layer initialization
This fix makes initialization of SDO client more robust, and
improves some comments for the initialization phase.
Due to time gap between sending an asynchronous frame and
internal state machine change of the SDO sequence layer
initialization it can happen that an initialization acknowledge
(InitAck, scon=1 / rcon=1) from SDO server is transmitted twice.
This should not count as error, and the connection should not be
closed. Instead, a timeout is issued in case the next correct
initialization frame from the server is not received.

Change-Id: I5272aaa863f3af50cf540098c635d91f1bbbeeb9