[FIX] Fix relict in padding area of asynchronous frames
The asynchronous Tx frame buffers are filled below the minimum Ethernet
frame size if not more data is provided by the higher layers, thus,
leaving the padding work to the Ethernet driver. If the Ethernet driver
uses the remaining frame buffer for padding, a relict from previously
sent frames could be used. Wireshark could misinterpret these SDO frames
and marks them as malformed packets.
This fix ensures that the padding area of the asynchronous frame buffer
is filled with zeros before forwarded to the Ethernet driver.

Change-Id: I1cb5d2996341196af9d22275388b12b5fcae3177