[FIX] Suppress compiler warnings for obsolete routines in Windows
Use of non-safe version of some standard routines such as open and
strncpy generates warning in Windows, as they are marked as
deprecated. The safe versions of these routines adds additional
memory checks which are not mandatory. Refer link for details
https://msdn.microsoft.com/en-us/library/8ef0s5kh(v=vs.120).aspx

The non-safe versions used in the stack are tested on different
platforms as well as the allocation and memory operations sizes
are fixed, which makes sure that non-safe routines will not lead
to any memory corruptions.

Considering above, this commit suppresses the compiler warnings
using the _CRT_SECURE_NO_WARNINGS compiler option without
affecting the code and performance.

Change-Id: Ia02c282cdd2c0d65b05f629827de9ca38d12c290
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>