[FIX] remove Veth code from edrv if Veth support is disabled
The openpowerlink driver build system doesn't take into account the
stack configuration (oplkcfg.h) to remove veth-*.c if Veth support
is disabled:
- veth-linuxkernel.c file from MODULE_SOURCE_FILES and
  HARDWARE_DRIVER_LINUXKERNEL_SOURCES
- veth-linuxuser.c from HARDWARE_DRIVER_LINUXUSER_SOURCES
- veth-ndisintemediate.c from HARDWARE_DRIVER_WINNDISIM_SOURCES
- veth-linuxpcie.c from HARDWARE_DRIVER_LINUXPCIE_SOURCES
- veth-generic.c from HARDWARE_DRIVER_OPENMAC_SOURCES

Thus, the Veth code is still present in generated edrv drivers.

Add a guard in each veth-*.c to hide the code if CONFIG_INCLUDE_VETH
is not set, like it is done in veth-linuxpcie.c.

Signed-off-by: Romain Naour <romain.naour@smile.fr>