[FIX] call pthread_setname_np() from the thread to be renamed
pthread_setname_np() must be used from the thread to be renamed
otherwise the new name will not be accepted when the openPOWERLINK
application is executed from a restricted user (i.e not root).

This is explained by the pthread_setname_np() implementation [1].

If we try to rename the thread calling pthread_setname_np(), then
"prctl (PR_SET_NAME, name)" will be used and the new name will be accepted.

But if we try to rename another thread, the PROCFS will be used
(/proc/self/task/<pid>/comm) and for some reason "comm" can't be written by
the user.

Make sure to rename a thread by calling pthread_setname_np() from the thread
to be renamed.

[1] https://sourceware.org/git/?p=glibc.git;a=blob;f=sysdeps/unix/sysv/linux/pthread_setname.c;h=6c5e1d6e485eaffed227156b6ccd75e54cc34a03;hb=fdfc9260b61d3d72541f18104d24c7bcb0ce5ca2#l31

Signed-off-by: Romain Naour <romain.naour@smile.fr>