[FIX] Fix wrong NMT state check in processFillTx()
In d0c47d3 the NMT state is checked to drop asynchronous Tx frames
silently if they are posted in a state where no POWERLINK
communication is established. The check does not consider a node
with auto-response capability running as MN, thus, frames are also
dropped in valid states forcing the node to hang in
kNmtMsPreOperational1 forever.

With this fix the NMT state is correctly interpreted so that frames
are dropped for NMT states below kNmtCsPreOperational1 and
kNmtMsPreOperational1.

Change-Id: Ida2a730bf1775909374a7738e7f7703003004526