[FIX] increase the i210 link up timeout
From time to time, the edrv i210 reach the link up timeout (3sec), thus the
openpowerlink stack fail to initialize correctly leading to the following
error:
oplk_create() with "Resource could not be created (Windows, PxROS, ...)" (0x0008)

By measuring the link up time, it appear that the nominal link up time is
"very" close to the timeout (~2.2sec to ~3.0+ sec).

Increase the timeout to 6sec in order to reach the timeout only if the
ethernet cable is really not connected.

Signed-off-by: Romain Naour <romain.naour@smile.fr>