[FIX] Solve NMT command truncation
This commit fixes the silent data size truncation in
nmtcnu_sendNmtRequestEx(). Now, the correct size of NmtCommandData is
received in handleObdRequestCmd() and used in nmtcnu_sendRequestEx(). If
the passed NmtCommandData size exceeds the limit of
C_MAX_NMT_CMD_DATA_SIZE, kErrorNmtInvalidParam is returned instead of
silently truncating the value.

Change-Id: Ie5eb05513d334e3b70d9a0f1e5fe0246757eb1dc