[TASK] Add pcap immediate mode support
In pcap immediate mode, packets are delivered to the stack instantly as
soon as they arrive. This allows shorter cycletimes in applications that
use pcap. The mode is supported by libpcap >= 1.5.0. The linux pcap
edrv falls back to the buffered mode if an older libpcap version is
detected.

Implements pull request #166 by sinseman44

Change-Id: I5a3fdea17859d534ee9ae610b675df1932ac0af9