[FIX] Avoid flood of error traces in pdou_cbNmtStateChange()
In debug mode the stack produces lots of error traces in
pdou_cbNmtStateChange() for Tx and Rx mapping objects that
do not exist in the OD. Since it is allowed to have less
mapping objects than possible the error can be ignored.
This fix ensures that error traces are not produced for
mapping objects that do not exist.

Change-Id: Ia0373c0970dfa1d2818e90036954849fe2985000