[FIX] NMTMNU: Handle missing PRC support of a node
The NMTMNU ignored the poll response chaining feature flag in
the ident response of a node. This could lead to a boot loop of
the corresponding node if configured to be used in
poll response chaining mode.
This fix changes the handling for such cases in the NMTMNU by
changing the node assignment temporary to asynchronous only. This
avoids the potential boot loop and enables the application and/or
the firmware manager to update the node's software revision that
supports poll response chaining.

Change-Id: Id70805f9f55338c1a443aeab378f4f22e9e1a1bf