[FIX] Fix object write and read API functions
Restore the previous behavior that the local object write and
read API functions are using obdu_writeEntry() and
obdu_readEntry() to access the OD. The remote object write and
read API functions are using obdu_writeEntryFromLe() and
obdu_readEntryToLe() if the local OD is accessed. This commit
reverts parts of 64c5a7f.

Change-Id: If45a1f015243d1c7069570491eb5ed36b8139cef