[FIX] Fix loss of frames on SDO client segmented WriteByIndex
If the SDO history buffer reaches its limits the SDO client
looses frames due to an update of the managing handle previous to
a blocked transmission. This unsuccessful transmission is never repeated
and therefore lost.
The solution is to update the managing handle only if the frame was
successfully sent or stored in the Tx history queue.
Since this change affects the SDO sequence layer, the same update
mechanism has to be applied to the SDO server.

Change-Id: Ifea590cf8e4528ee4e48ba5f10bdbe969d7442e3