[FIX] Fix NIOS II store-restore default settings
The configuration storage feature can only be activated if the
host cpu board and the NIOS II BSP supports it. Per default,
this feature is switched off in the application, thus it also
has to be switched off in the hostif-library because the BSP
does not support it. This causes a fatal error at the host
while reading the CFI-flash memory.

To enable the configuration storage feature, set both:
* -DCONFIG_APP_STORE_RESTORE
  in apps/demo_cn_embedded/build/altera-nios2/app.settings
* -DCONFIG_INCLUDE_STORE_RESTORE
  in stack/build/altera-nios2/liboplkcnapp-hostif/lib.settings

Change-Id: I4253dcfac6099e180aa84f0edbdb06fbdd0bdfe0