[FIX] Correct openCONFORMANCE SoC error detection behavior of CN
For consecutive SoC losses, which means no SoC is received for at least
2 cycles in a row, only SoC timeouts are used to trigger a correspondig
error reporting and counting in order to avoid multiple counts for the
same SoC loss. In this phase, a logical detection of a lost SoC frame
(per frame sequece check) can be used only once, which might speed up
the detection for one SoC loss. As soon as an SoC is detected again,
any SoC loss detected logically or by timeout will be used to count
and report an SoC loss.

Fixes openCONFORMANCE testcases:
* Testcase_3_3_1_1_T1
* Testcase_3_3_1_1_T2
* Testcase_3_3_1_1_T7

Change-Id: I526e870abb90cff8c720d3a6558a57785aa69c28